% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/non_informative_var.R
\name{is.not.informative.variable}
\alias{is.not.informative.variable}
\title{Is a vector an non informative variable}
\usage{
is.not.informative.variable(x, includeNA = T)
}
\arguments{
\item{x}{a vector of numeric or factor.}

\item{includeNA}{a boolean. TRUE to include NA value as a factor level.}
}
\description{
This function determines if a given vector of numeric or factor is a non informative variable or not.
}
\examples{
data(iris)
is.not.informative.variable(iris$Sepal.Length)

}
