% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkspotterComplete.R
\name{linkspotterComplete}
\alias{linkspotterComplete}
\title{Linspotter}
\usage{
linkspotterComplete(dataset, corMethods = c("pearson", "spearman", "kendall",
  "mic", "MaxNormMutInfo"), defaultMinCor = 0.3,
  defaultCorMethod = "MaxNormMutInfo", clusteringCorMethod = NULL,
  nbCluster = 1:9, printInfo = T)
}
\arguments{
\item{dataset}{the dataframe which variables bivariate correlations are to be analyzed.}

\item{corMethods}{a vector of correlation coefficients to compute. The available coefficient are the following : \code{c("pearson","spearman","kendall","mic","distCor","MaxNormMutInfo")}. It is not case sensitive and still work if only the beginning of the word is put (e.g. \code{pears}).}

\item{defaultMinCor}{a double between 0 and 1. It is the minimal correlation absolute value to consider for the first graph plot.}

\item{defaultCorMethod}{a string. One of "pearson","spearman","kendall","mic", "distCor" or "MaxNormMutInfo". It is the correlation coefficient to consider for the first graph plot.}

\item{clusteringCorMethod}{a string. One of "pearson","spearman","kendall","mic", "distCor" or "MaxNormMutInfo". It is the correlation coefficient to consider for the variables clustering.}

\item{nbCluster}{an integer. It is the number of clusters to compute.}

\item{printInfo}{a boolean indicating whether to print on the console some information about the dataset and the estimated computation time.}
}
\value{
a list containing all the material enabling to analyze correlations:
\itemize{
  \item \code{computationTime}: a string
  \item \code{run_it}: a shiny.appobj object enable to deploy instantly the user interface for a customizable visualization.
  \item \code{dataset}: the initial dataset
  \item \code{corDF}: a the correlation data.frame including values for all coefficients
  \item \code{corMatrices}: a list of correlation matrices
  \item \code{corGroups}: data.frame a data.frame list
  \item \code{clusteringCorMethod}: a character
  \item \code{defaultMinCor}: a numeric
  \item \code{defaultCorMethod}: a string
  \item \code{corMethods}: vector of strings
}
}
\description{
Computation of correlation matrices, variable clustering and the customizable user inferface to visualize them using a graph and visualize variables distributions and cross plots.
}
