% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkspotterOnFile.R
\name{linkspotterOnFile}
\alias{linkspotterOnFile}
\title{Linspotter / linkspotterOnFile a function to load and proccess Linkspotter on a file}
\usage{
linkspotterOnFile(file, header, sep, quote, corMethods = c("pearson",
  "spearman", "kendall", "mic", "MaxNMI"), defaultMinCor = 0.3,
  defaultCorMethod = corMethods[length(corMethods)],
  clusteringCorMethod = corMethods[length(corMethods)], nbCluster = 1:9,
  printInfo = T, appTitle = "Linkspotter", ...)
}
\arguments{
\item{file}{the file containing a structured dataset which the bivariate correlations are to be analyzed.}

\item{header}{a boolean.TRUE if the file contains a header (similar to read.table)}

\item{sep}{the separation character used in the file (similar to read.table)}

\item{quote}{the quote character used in the file (similar to read.table)}

\item{corMethods}{a vector of correlation coefficients to compute. The available coefficients are the following : \code{c("pearson","spearman","kendall","mic","distCor","MaxNMI")}. It is not case sensitive and still work if only the beginning of the word is put (e.g. \code{pears}).}

\item{defaultMinCor}{a double between 0 and 1. It is the minimal correlation absolute value to consider for the first graph plot.}

\item{defaultCorMethod}{a string. One of "pearson","spearman","kendall","mic", "distCor" or "MaxNMI". It is the correlation coefficient to consider for the first graph plot.}

\item{clusteringCorMethod}{a string. One of "pearson","spearman","kendall","mic", "distCor" or "MaxNMI". It is the correlation coefficient to consider for the variables clustering.}

\item{nbCluster}{an integer. It is the number of clusters to compute.}

\item{printInfo}{a boolean indicating whether to print on the console some information about the dataset and the estimated computation time.}

\item{appTitle}{a string taken as the title of the user interface.}

\item{...}{Further arguments to be passed to the used read.table function.}
}
\value{
a list containing all the material enabling to analyze correlations:
\itemize{
  \item{\code{computationTime}: a string}
  \item{\code{run_it}: a shiny.appobj object enable to deploy instantly the user interface for a customizable visualization.}
  \item{\code{dataset}: the initial dataset}
  \item{\code{corDF}: a the correlation data.frame including values for all coefficients}
  \item{\code{corMatrices}: a list of correlation matrices}
  \item{\code{corGroups}: data.frame a data.frame list}
  \item{\code{clusteringCorMethod}: a character}
  \item{\code{defaultMinCor}: a numeric}
  \item{\code{defaultCorMethod}: a string}
  \item{\code{corMethods}: vector of strings}
}
}
\description{
Computation of correlation matrices, variable clustering and the customizable user inferface to visualize them using a graph together with variables distributions and cross plots.
}
\examples{
\dontrun{

# run linkspotter on iris example data
data(iris)
lsOutputIrisFromFile<-linkspotterOnFile("iris.csv")

# launch the UI
lsOutputIrisFromFile$run_it

}

}
