% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxNMI.R
\name{maxNMI}
\alias{maxNMI}
\title{Linspotter / Maximal Normalized Mutual Information (MaxNMI) function}
\usage{
maxNMI(x, y)
}
\arguments{
\item{x}{a vector of numeric or factor.}

\item{y}{a vector of numeric or factor.}
}
\value{
a double between 0 and 1 corresponding to the MaxNMI.
}
\description{
Calculate the MaxNMI relationship measurement
}
\examples{
\dontrun{

# calculate a correlation dataframe
data(iris)
maxNMI(iris$Sepal.Length,iris$Sepal.Width)

}

}
