% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxNMI_numfact.R
\name{maxNMI_numfact}
\alias{maxNMI_numfact}
\title{Linspotter / Maximal Normalized Mutual Information (MaxNMI) function for a numeric vs a factor variable}
\usage{
maxNMI_numfact(continuousY, factorX, includeNA = T)
}
\arguments{
\item{continuousY}{a vector of numeric.}

\item{factorX}{a vector of factor.}

\item{includeNA}{a boolean. TRUE to include NA value as a factor level.}
}
\value{
a double between 0 and 1 corresponding to the MaxNMI.
}
\description{
Calculate the MaxNMI relationship measurement for a numeric vs a factor variable
}
\examples{
\dontrun{

# calculate a correlation dataframe
data(iris)
maxNMI_numfact(continuousY=iris$Sepal.Length,factorX=iris$Species)

}

#@export

}
