% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxNMI_numnum.R
\name{maxNMI_numnum}
\alias{maxNMI_numnum}
\title{Linspotter / Maximal Normalized Mutual Information (MaxNMI) function for a couple of numeric variable}
\usage{
maxNMI_numnum(continuousX, continuousY, maxNbBins = NA)
}
\arguments{
\item{continuousX}{a vector of numeric.}

\item{continuousY}{a vector of numeric.}

\item{maxNbBins}{an integer corresponding to the number of bin limitation, maxNbBins=100 by default.}
}
\value{
a double between 0 and 1 corresponding to the MaxNMI.
}
\description{
Calculate the MaxNMI relationship measurement between to numeric variables
}
\examples{
\dontrun{

# calculate a correlation dataframe
data(iris)
maxNMI_numnum(iris$Sepal.Length,iris$Sepal.Width)

}

#@export

}
