% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corCouplesToMatrix.R
\name{corCouplesToMatrix}
\alias{corCouplesToMatrix}
\title{Transform a correlation dataframe into a correlation matrix}
\usage{
corCouplesToMatrix(x1_x2_val)
}
\arguments{
\item{x1_x2_val}{a specific dataframe containing correlations values resulting from the function multiBivariateCorrelation() and containing only one coeffecient type.}
}
\value{
a dataframe corresponding to a correlation matrix.
}
\description{
Transform a 2 column correlation dataframe into a correlation matrix
}
\examples{
\dontrun{

# calculate a correlation dataframe
data(iris)
corDF=multiBivariateCorrelation(mixedData = iris, corMethods = "MaxNMI")
corMatrix=corCouplesToMatrix(x1_x2_val = corDF[,c('X1','X2',"MaxNMI")])
print(corMatrix)
}

}
