% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkspotterUI.R
\name{linkspotterUI}
\alias{linkspotterUI}
\title{Linspotter user interface runner}
\usage{
linkspotterUI(dataset, corDF, variablesClustering = NULL,
  defaultMinCor = 0.3, appTitle = "Linkspotter")
}
\arguments{
\item{dataset}{the dataframe which variables bivariate correlations are contained in corDF}

\item{corDF}{a specific dataframe containing correlations values resulting from the function multiBivariateCorrelation()}

\item{variablesClustering}{a specific dataframe containing the output of the variable clustering resulting from the function clusterVariables()}

\item{defaultMinCor}{a double between 0 and 1. It is the minimal correlation absolute value to consider for the first graph plot.}

\item{appTitle}{a string taken as the title of the user interface.}
}
\value{
a shiny.appobj object enable to deploy instantly the user interface for a customizable visualization.
}
\description{
Run the linkSpotter user interface
}
\examples{
\dontrun{

# calculate a correlation dataframe
data(iris)
corDF=multiBivariateCorrelation(mixedData = iris, corMethods = "MaxNMI")
corMatrix=matrixOfValuesOfAllCouples(x1_x2_val = corDF[,c('X1','X2',"MaxNMI")])
corGroups=clusterVariables(correlation_matrix = corMatrix, nbCluster = 3)

# launch the UI
linkspotterUI(dataset=iris,
corDF=corDF,
variablesClustering=corGroups,
defaultMinCor=0.3,
appTitle="Linkspotter on iris data")

}

}
